Complete users api

+++ src/controllers/UserController.js
@@ -1,7 +1,16 @@
 const repository = require('../repositories/userRepo');
+const auth = require('../middlewares/auth');
+const jwt = require('../utils/jwt');
+const { MODULES } = require('../../config');
 
 const updateUser = async (req, res) => {
   try {
+    const { token } = req.headers;
+    const permission = await auth('update', MODULES.users, token);
+
+    if (!permission) {
+      return res.send('Permission denied!');
+    }
     const { id } = req.params;
     const foundUser = await repository.getById(id);
     if (foundUser) {
@@ -17,11 +26,17 @@ const updateUser = async (req, res) => {
 
 const deleteUser = async (req, res) => {
   try {
+    const { token } = req.headers;
+    const permission = await auth('delete', MODULES.users, token);
+
+    if (!permission) {
+      return res.send('Permission denied!');
+    }
     const { id } = req.params;
     const foundUser = await repository.getById(id);
     if (foundUser) {
       const user = await repository.deleteUser(id);
-      res.send(user);
+      return res.send('User successfully deleted!');
     }
     res.status(404).send('User with this ID not found!');
   } catch (error) {
@@ -31,6 +46,13 @@ const deleteUser = async (req, res) => {
 
 const getUsers = async (req, res) => {
   try {
+    const { token } = req.headers;
+    const permission = await auth('read', MODULES.users, token);
+    console.log(permission);
+
+    if (!permission) {
+      return res.send('Permission denied!');
+    }
     const users = await repository.getUsers();
     res.send(users);
   } catch (error) {
@@ -40,6 +62,12 @@ const getUsers = async (req, res) => {
 
 const getUser = async (req, res) => {
   try {
+    const { token } = req.headers;
+    const permission = await auth('read', MODULES.users, token);
+
+    if (!permission) {
+      return res.send('Permission denied!');
+    }
     const { id } = req.params;
     const user = await repository.getById(id);
     res.send(user);
@@ -50,7 +78,18 @@ const getUser = async (req, res) => {
 
 const insertUser = async (req, res) => {
   try {
-    const user = await repository.insertUser(req.body);
+    const { token } = req.headers;
+    const permission = await auth('insert', MODULES.users, token);
+
+    if (!permission) {
+      return res.send('Permission denied!');
+    }
+    const phone_number = req.body.phone_number.replace('+', '');
+    const { full_name, password } = req.body;
+
+    const payload = jwt.verify(token);
+
+    const user = await repository.insertUser(full_name, phone_number, password, payload.userId);
     res.send(user);
   } catch (error) {
     throw Error(`users controller [insertUser]:${error}`);
